Start a separate file for CSRF tests
